Added Pods and Search functionality
Added Firebase pods
Added ability to search
Added tv show functionality

+++ MoviesWithFriends/Models/MediaInfo.swift
@@ -12,7 +12,7 @@ protocol MediaInfo: MediaDisplayable {
     var releaseDate: String { get }
     var rating: String? { get }
     var reviewScore: Double? { get }
-    var runtime: Double? { get }
+    var runtime: Int? { get }
     var genres: [Genre] { get }
     var credits: Credits { get }
     var videos: Videos { get }
